fix Post vote toggling issue
Toggled attrs are now always false in Post DB, but updated locally
on a per user basis upon Group component load. Vote counts are updated
in DB, toggled state stays false. In changing the toggled state to be
based on user votes, I ran into an issue with fetching votes within
PostService getPostsByGroupName(). The problem was a nested
subscription to feeds, inside which getPost() was being called. This
ended up breaking the toggled state load based on votes in Group
Component. The solution was to change the service method to only fetch
the IDs, then forEach(getPost()) inside the Group component itself.
Users are created with an empty vote array.