fix simultaneous multi-users voting bug
If two users were logged in at the same time, and voted on a post
at the same time (i.e. in one session without refreshing their apps),
the vote count in the DB would not properly change. This was due to
the fact that the DB post entry was being assigned to that of the
locally changed post. To fix, on upvote/downvoteClick, we fetch current
value of vote counts, and increment/decrement from there.