add sort posts functionality on Group (top, newest, oldest)
Selecting a sort radio button will sort the component's post list
(and live re-render the posts). If another user makes a vote that
changes the DB order of top posts by vote or submits a new post,
and the current user refreshPosts(), the sort will be reapplied
after updating this.posts as a part of refreshPosts().